Updated volunteer 'access' roles
Implemented 'isActive' to determine whether a user is an active volunteer. This function can be changed only in firebase and viewed there since it is admin information. Automatically becomes active upon a user signing up. Implemented automatic access roles such as secchi sim, secchi videos, and quizzes, meanwhile the other roles are set to false. The 'accessRoles' functions cannot be changed if desired for users, only through firebase (for admin).